Sources/Version2/Diaspora-ImageProcessing/requirements.txt
# ============================================================
# AI 기반 한인 디아스포라 기록 유산 디지털화 시스템
# 이미지 처리 모듈 - 확정 패키지 버전 목록
# 
# Python: 3.10.11
# CUDA: 12.1
# ============================================================

# ── Core ────────────────────────────────────────────────────
torch==2.1.2
torchvision==0.16.2
numpy==1.26.4
Pillow>=10.0.0
opencv-python>=4.8.0

# ── Super Resolution (Real-ESRGAN) ──────────────────────────
# 설치 시 반드시 --no-deps 사용:
#   pip install basicsr==1.4.2 --no-deps
#   pip install realesrgan==0.3.0 --no-deps
basicsr==1.4.2
realesrgan==0.3.0

# ── Face Restoration (GFPGAN) ───────────────────────────────
gfpgan>=1.3.8
facexlib>=0.3.0

# ── Colorization (DDColor) ──────────────────────────────────
# pip install 불가 — GitHub에서 직접 설치 필요:
#   git clone https://github.com/piddnad/DDColor.git ddcolor_src
#   cd ddcolor_src && pip install -e . --no-build-isolation && cd ..
# 설치 후 numpy/basicsr 재고정 필수 (위 버전으로)
# HuggingFace 모델: piddnad/ddcolor_modelscope (자동 다운로드)

# ── Image Analysis (BLIP, CLIP) ─────────────────────────────
# transformers 5.x 사용 불가 (torch 2.4+ 요구)
transformers==4.37.2
tokenizers==0.15.2
huggingface-hub==0.20.3

# ── Face Recognition ────────────────────────────────────────
face_recognition>=1.3.0
dlib>=19.24.0

# ── Utilities ───────────────────────────────────────────────
tqdm>=4.65.0
requests>=2.31.0
pyyaml>=6.0
colorama>=0.4.6
scikit-image>=0.21.0
lpips>=0.1.4
